removed flipper
